Updated check in and out with api calls
