+++ client/src/actions/types.js
@@ -8,3 +8,4 @@ export const GET_ITEMS = "GET_ITEMS";
 export const TRANSACTION_LOADING = "ITEM_LOADING";
 export const SET_CURRENT_TRANSACTION = "SET_CURRENT_ITEM";
 export const GET_TRANSACTIONS = "GET_ITEMS";
+
