+++ client/src/reducers/itemReducers.js
@@ -5,7 +5,7 @@ import {
     UPDATE_ITEM,
     CREATE_ITEM
 } from "../actions/itemActions";
-//const isEmpty = require("is-empty");
+const isEmpty = require("is-empty");
 
 export default function (
     state = {
@@ -33,19 +33,3 @@ export default function (
     }
 }
 
-// export default function (state = initialState, action) {
-//     switch (action.type) {
-//         case GET_USERS:
-//             return {
-//                 ...state,
-//                 users: action.payload
-//             };
-//         case USER_LOADING:
-//             return {
-//                 ...state,
-//                 loading: true
-//             };
-//         default:
-//             return state;
-//     }
-// }
