+++ routes/api.js
@@ -20,10 +20,16 @@ const validateResetInput = require("../validation/reset");
 const User = require('../models/User');
 const Item = require('../models/Item');
 const Transaction = require("../models/Transaction");
+const { Types } = require('mongoose');
 
-// Transaction Routes
+/////////////////////////////////////////////// Transaction Routes /////////////////////////////////////////////////
 router.get('/transactions', (req, res) => {
-    Transaction.find({})
+    Transaction.aggregate([
+        { $lookup: { from: "items", localField: "item_id", foreignField: "_id", as: "item" } },
+        { $lookup: { from: "users", localField: "user_id", foreignField: "_id", as: "user" } },
+        { $unwind: "$item" },
+        { $unwind: "$user" }
+    ])
         .then((data) => {
             console.log('Data: ', data);
             res.json(data);
@@ -45,19 +51,69 @@ router.post('/transactions', (req, res) => {
 })
 
 router.put('/transactions/:id', (req, res) => {
-    Transaction.findByIdAndUpdate(req.params.id, req.body, { new: true })
-        .then(transaction => {
-            if (!transaction)
-                return res.status(404).send({ message: "Transaction not found with id " + req.params.id });
-            res.send(item)
-        }).catch(err => {
-            if (err.kind === 'ObjectId')
-                return res.status(404).send({ message: "Item not found with id " + req.params.id });
-            return res.status(500).send({ message: "Error updating item with id " + req.params.id });
+    Transaction.findOne({ _id: req.params.id }, function (err, transaction) {
+        transaction.checkedInDate = req.body.checkedInDate;
+        transaction.save();
+    })
+        .then((data) => {
+            res.json(data);
         })
+        .catch((error) => {
+            console.log('error: ', error);
+        });
 })
 
-// Item Routes
+router.get('/transactions/findbyuser/:id', (req, res) => {
+    console.log(req.params);
+    Transaction.aggregate([
+        {
+            "$match": {
+                "$and": [
+                    { "user_id": Types.ObjectId(req.params.id) },
+                    { "checkedInDate": null },
+                ]
+            }
+        },
+        { $lookup: { from: "items", localField: "item_id", foreignField: "_id", as: "item" } },
+        { $lookup: { from: "users", localField: "user_id", foreignField: "_id", as: "user" } },
+        { $unwind: "$item" },
+        { $unwind: "$user" }
+    ])
+        .then((data) => {
+            console.log('Data: ', data);
+            res.json(data);
+        })
+        .catch((error) => {
+            console.log('error: ', error);
+        });
+});
+
+router.get('/transactions/findbyuserall/:id', (req, res) => {
+    console.log(req.params);
+    Transaction.aggregate([
+        {
+            "$match": {
+                "$and": [
+                    { "user_id": Types.ObjectId(req.params.id) },
+                ]
+            }
+        },
+        { $lookup: { from: "items", localField: "item_id", foreignField: "_id", as: "item" } },
+        { $lookup: { from: "users", localField: "user_id", foreignField: "_id", as: "user" } },
+        { $unwind: "$item" },
+        { $unwind: "$user" }
+    ])
+        .then((data) => {
+            console.log('Data: ', data);
+            res.json(data);
+        })
+        .catch((error) => {
+            console.log('error: ', error);
+        });
+});
+
+
+//.///////////////////////////////// Item Routes //////////////////////////////////////////////////////
 router.get('/items', (req, res) => {
     Item.find({})
         .then((data) => {
@@ -68,6 +124,50 @@ router.get('/items', (req, res) => {
             console.log('error: ', error);
         });
 });
+
+router.get('/items/available', (req, res) => {
+    Item.aggregate([
+        {
+            $lookup:
+            {
+                from: "transactions",
+                localField: "_id",
+                foreignField: "item_id",
+                as: "transactions",
+            }
+        },
+        {
+            $project: {
+                id : "$_id",
+                transactions: "$transactions",
+                courses: "$courses",
+                name: "$name",
+                iid: "$iid",
+                serial: "$serial",
+                category: "$category",
+                notes: "$notes",
+                brand: "$brand",
+                activeTransaction: {
+                    $filter : {
+                        input: "$transactions",
+                        as : "activeTransaction",
+                        cond : {
+                            $eq: ["$$activeTransaction.checkedInDate",null]
+                        }
+                    }
+                }
+            }
+        },
+    ])
+        .then((data) => {
+            console.log('Data: ', data);
+            res.json(data);
+        })
+        .catch((error) => {
+            console.log('error: ', error);
+        });
+});
+
 
 router.post('/items', (req, res) => {
     const newItem = new Item(req.body);
@@ -99,6 +199,17 @@ router.get('/users', (req, res) => {
     User.find({})
         .then((data) => {
 
+            data.forEach(user => user.password = null);
+            res.json(data);
+        })
+        .catch((error) => {
+            console.log('error: ', error);
+        });
+});
+
+router.get('/users/:ucode', (req, res) => {
+    User.findOne({ userCode: req.params.ucode })
+        .then((data) => {
             res.json(data);
         })
         .catch((error) => {
@@ -127,8 +238,9 @@ router.post('/users', (req, res) => {
     })
 })
 
-router.put('/users/:id', (req, res) => {
-    User.findByIdAndUpdate(req.params.id, req.body, { new: true })
+router.put('/users/', (req, res) => {
+    console.log(req);
+    User.findByIdAndUpdate(req.body._id, req.body, { new: true })
         .then(user => {
             if (!user)
                 return res.status(404).send({ message: "User not found with id " + req.params.id });
@@ -148,6 +260,8 @@ router.delete('/users/:id', (req, res) => {
             return res.status(500).send({ message: "Error updating user with id " + req.params.id });
         })
 })
+
+
 
 // @route POST api/users/register
 // @desc Register user
